+++ pkg/template/template_test.go
@@ -341,3 +341,78 @@ func TestExecute(t *testing.T) {
 		})
 	}
 }
+
+func TestTruncateMultiline(t *testing.T) {
+	type args struct {
+		max int
+		s   string
+	}
+	tests := []struct {
+		name string
+		args args
+		want string
+	}{
+		{
+			name: "exactly minimum width",
+			args: args{
+				max: 5,
+				s:   "short",
+			},
+			want: "short",
+		},
+		{
+			name: "exactly minimum width with new line",
+			args: args{
+				max: 5,
+				s:   "short\n",
+			},
+			want: "sh...",
+		},
+		{
+			name: "less than minimum width",
+			args: args{
+				max: 4,
+				s:   "short",
+			},
+			want: "shor",
+		},
+		{
+			name: "less than minimum width with new line",
+			args: args{
+				max: 4,
+				s:   "short\n",
+			},
+			want: "shor",
+		},
+		{
+			name: "first line of multiple is short enough",
+			args: args{
+				max: 80,
+				s:   "short\n\nthis is a new line",
+			},
+			want: "short...",
+		},
+		{
+			name: "using Windows line endings",
+			args: args{
+				max: 80,
+				s:   "short\r\n\r\nthis is a new line",
+			},
+			want: "short...",
+		},
+		{
+			name: "using older MacOS line endings",
+			args: args{
+				max: 80,
+				s:   "short\r\rthis is a new line",
+			},
+			want: "short...",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := truncateMultiline(tt.args.max, tt.args.s)
+			assert.Equal(t, tt.want, got)
+		})
+	}
+}
