+++ pkg/text/text.go
@@ -3,16 +3,12 @@ package text
 
 import (
 	"fmt"
-	"net/url"
 	"regexp"
 	"strings"
 	"time"
-	"unicode"
 
 	"github.com/muesli/reflow/ansi"
 	"github.com/muesli/reflow/truncate"
-	"golang.org/x/text/cases"
-	"golang.org/x/text/language"
 )
 
 const (
@@ -21,52 +17,14 @@ const (
 )
 
 var indentRE = regexp.MustCompile(`(?m)^`)
-var whitespaceRE = regexp.MustCompile(`\s+`)
 
-// Indent returns a copy of the string s with indent prefixed to it, will indent each line
-// in the string.
+// Indent returns a copy of the string s with indent prefixed to it, will apply indent
+// to each line of the string.
 func Indent(s, indent string) string {
 	if len(strings.TrimSpace(s)) == 0 {
 		return s
 	}
 	return indentRE.ReplaceAllLiteralString(s, indent)
-}
-
-// CamelToKebab returns a copy of the string s that is converted from camel case form to '-' separated form.
-func CamelToKebab(s string) string {
-	var output []rune
-	var segment []rune
-	for _, r := range s {
-		if !unicode.IsLower(r) && string(r) != "-" && !unicode.IsNumber(r) {
-			output = addSegment(output, segment)
-			segment = nil
-		}
-		segment = append(segment, unicode.ToLower(r))
-	}
-	output = addSegment(output, segment)
-	return string(output)
-}
-
-func addSegment(inrune, segment []rune) []rune {
-	if len(segment) == 0 {
-		return inrune
-	}
-	if len(inrune) != 0 {
-		inrune = append(inrune, '-')
-	}
-	inrune = append(inrune, segment...)
-	return inrune
-}
-
-// Title returns a copy of the string s with all Unicode letters that begin words mapped to their Unicode title case.
-func Title(s string) string {
-	c := cases.Title(language.English)
-	return c.String(s)
-}
-
-// RemoveExcessiveWhitespace returns a copy of the string s with excessive whitespace removed.
-func RemoveExcessiveWhitespace(s string) string {
-	return whitespaceRE.ReplaceAllString(strings.TrimSpace(s), " ")
 }
 
 // DisplayWidth calculates what the rendered width of string s will be.
@@ -91,16 +49,6 @@ func Truncate(maxWidth int, s string) string {
 	return r
 }
 
-// TruncateMultiline returns a copy of the string s that has been shortened to fit the maximum
-// display width. If string s has multiple lines the first line will be shortened and all others
-// removed.
-func TruncateMultiline(maxWidth int, s string) string {
-	if i := strings.IndexAny(s, "\r\n"); i >= 0 {
-		s = s[:i] + ellipsis
-	}
-	return Truncate(maxWidth, s)
-}
-
 // Pluralize returns a concatenated string with num and the plural form of thing if necessary.
 func Pluralize(num int, thing string) string {
 	if num == 1 {
@@ -113,9 +61,9 @@ func fmtDuration(amount int, unit string) string {
 	return fmt.Sprintf("about %s ago", Pluralize(amount, unit))
 }
 
-// FuzzyAgo returns a human readable string of the time duration between a and b that is estimated
+// RelativeTimeAgo returns a human readable string of the time duration between a and b that is estimated
 // to the nearest unit of time.
-func FuzzyAgo(a, b time.Time) string {
+func RelativeTimeAgo(a, b time.Time) string {
 	ago := a.Sub(b)
 
 	if ago < time.Minute {
@@ -135,44 +83,4 @@ func FuzzyAgo(a, b time.Time) string {
 	}
 
 	return fmtDuration(int(ago.Hours()/24/365), "year")
-}
-
-// FuzzyAgoAbbr is an abbreviated version of FuzzyAgo. It returns a human readable string of the
-// time duration between a and b that is estimated to the nearest unit of time.
-func FuzzyAgoAbbr(a, b time.Time) string {
-	ago := a.Sub(b)
-
-	if ago < time.Hour {
-		return fmt.Sprintf("%d%s", int(ago.Minutes()), "m")
-	}
-	if ago < 24*time.Hour {
-		return fmt.Sprintf("%d%s", int(ago.Hours()), "h")
-	}
-	if ago < 30*24*time.Hour {
-		return fmt.Sprintf("%d%s", int(ago.Hours())/24, "d")
-	}
-
-	return b.Format("Jan _2, 2006")
-}
-
-// Humanize returns a copy of the string s that replaces all instance of '-' and '_' with spaces.
-func Humanize(s string) string {
-	replace := "_-"
-	h := func(r rune) rune {
-		if strings.ContainsRune(replace, r) {
-			return ' '
-		}
-		return r
-	}
-	return strings.Map(h, s)
-}
-
-// DisplayURL returns a copy of the string urlStr removing everything except the hostname and path.
-// If there is an error parsing urlStr then urlStr is returned without modification.
-func DisplayURL(urlStr string) string {
-	u, err := url.Parse(urlStr)
-	if err != nil {
-		return urlStr
-	}
-	return u.Hostname() + u.Path
 }
